moved to formatUtilities
