+++ src/common/Calculations.js
@@ -33,63 +33,6 @@ const CalculateTaxCollected = (
   taxRate = RatesAndFees.TransientTaxRate
 ) => roundCurrency(netRoomRentalCollections * taxRate);
 
-const ParseAmountToFloat = amount => {
-  return parseFloat(amount.toString().replace(/,/g, ""));
-};
-const FormatNumber = n => {
-  // format number 1000000 to 1,234,567
-  return n
-    .toString()
-    .replace(/^0+/, "")
-    .replace(/\D/g, "")
-    .replace(/\B(?=(\d{3})+(?!\d))/g, ",");
-};
-const PreserveDecimalFormatNumber = n => {
-  return n.toString().indexOf(",") >= 0
-    ? n
-    : n.toString().replace(/\B(?=(\d{3})+(?!\d))/g, ",");
-};
-const FormattedAmount = (value, isNegativeValue) => {
-  let fieldValue = 0;
-  if (value.trim() === "") {
-    return [0, fieldValue];
-  }
-
-  if (value.indexOf(".") >= 0) {
-    var decimal_pos = value.indexOf(".");
-    var left_side = value.substring(0, decimal_pos);
-    var right_side = value.substring(decimal_pos);
-    if (left_side.length > 0) {
-      value =
-        FormatNumber(left_side) +
-        "." +
-        FormatNumber(right_side).substring(0, 2);
-
-      fieldValue = isNegativeValue
-        ? -ParseAmountToFloat(value)
-        : ParseAmountToFloat(value);
-    } else {
-      var rightSideLength = right_side.length;
-      var rightFormatedValue = "." + FormatNumber(right_side).substring(0, 2);
-      value = rightSideLength === 1 ? value : rightFormatedValue;
-      fieldValue =
-        rightSideLength === 1
-          ? 0
-          : isNegativeValue
-          ? -ParseAmountToFloat(rightFormatedValue)
-          : ParseAmountToFloat(rightFormatedValue);
-    }
-
-    fieldValue = ParseAmountToFloat(fieldValue);
-  } else {
-    value = value === 0 ? 0 : FormatNumber(value);
-    fieldValue = isNegativeValue
-      ? -ParseAmountToFloat(value)
-      : ParseAmountToFloat(value);
-  }
-  return [value, fieldValue];
-};
-
 /**
  * Get a list of calculations based on given Transient Tax Data
  * @param {object} fields the required input fields to calculate totals for transient tax. Fields are grouped by their form name.
@@ -190,7 +133,5 @@ export {
   CalculateTaxCollected,
   GetCalculatedTotals,
   GetTotalsForMonth,
-  SumTotals,
-  PreserveDecimalFormatNumber,
-  FormattedAmount
+  SumTotals
 };
