+++ src/common/FormatUtilities.js
@@ -18,4 +18,84 @@ const FormatCurrency = dollarAmount => currencyFormatter.format(dollarAmount);
  */
 const FormatPercentage = percentageAsDecimal => `${percentageAsDecimal * 100}%`;
 
-export { FormatCurrency, FormatPercentage };
+/**
+ * Parse string to float
+ * @param {string} amount
+ */
+const ParseAmountToFloat = amount => {
+  return parseFloat(amount.toString().replace(/,/g, ""));
+};
+/**
+ * Format enter amount to currency string
+ * @param {string} n
+ */
+const FormatNumber = n => {
+  // format number 1000000 to 1,234,567
+  return n
+    .toString()
+    .replace(/^0+/, "")
+    .replace(/\D/g, "")
+    .replace(/\B(?=(\d{3})+(?!\d))/g, ",");
+};
+/**
+ * preserve currency string with deciaml when previous button is clicked
+ * @param {string} n
+ */
+const PreserveDecimalFormatNumber = n => {
+  return n.toString().indexOf(",") >= 0
+    ? n
+    : n.toString().replace(/\B(?=(\d{3})+(?!\d))/g, ",");
+};
+
+/**
+ * Format value to curreny text format 
+ * @param {string} value
+ * @param {boolean} isNegativeValue
+ */
+const FormattedAmount = (value, isNegativeValue) => {
+  let fieldValue = 0;
+  if (value.trim() === "") {
+    return [0, fieldValue];
+  }
+
+  if (value.indexOf(".") >= 0) {
+    var decimal_pos = value.indexOf(".");
+    var left_side = value.substring(0, decimal_pos);
+    var right_side = value.substring(decimal_pos);
+    if (left_side.length > 0) {
+      value =
+        FormatNumber(left_side) +
+        "." +
+        FormatNumber(right_side).substring(0, 2);
+
+      fieldValue = isNegativeValue
+        ? -ParseAmountToFloat(value)
+        : ParseAmountToFloat(value);
+    } else {
+      var rightSideLength = right_side.length;
+      var rightFormatedValue = "." + FormatNumber(right_side).substring(0, 2);
+      value = rightSideLength === 1 ? value : rightFormatedValue;
+      fieldValue =
+        rightSideLength === 1
+          ? 0
+          : isNegativeValue
+          ? -ParseAmountToFloat(rightFormatedValue)
+          : ParseAmountToFloat(rightFormatedValue);
+    }
+
+    fieldValue = ParseAmountToFloat(fieldValue);
+  } else {
+    value = value === 0 ? 0 : FormatNumber(value);
+    fieldValue = isNegativeValue
+      ? -ParseAmountToFloat(value)
+      : ParseAmountToFloat(value);
+  }
+  return [value, fieldValue];
+};
+
+export {
+  FormatCurrency,
+  FormatPercentage,
+  PreserveDecimalFormatNumber,
+  FormattedAmount
+};
